+++ flow-typed/interfaces/ads-plugin-controller.js
@@ -1,0 +1,10 @@
+//@flow
+
+declare interface IAdsPluginController {
+  skipAd(): void;
+  playAdNow(adTagUrl: string): void;
+  onPlaybackEnded(): Promise<void>;
+  +active: boolean;
+  +done: boolean;
+  +name: string;
+}